==> 7_statistical_analysis.ipynb <==
import pandas as pd
from scipy.stats import friedmanchisquare
import scikit_posthocs as sp
import numpy as np
import seaborn as sns
import pingouin as pg

cv = pd.read_csv('cv_results.csv')

# replace abbreviations in model column of cv
cv['model'] = cv['model'].replace({
    'GFN': 'TGFNN'
})

cv['dataset'] = cv['dataset'].replace(
    {
        '30_phenotypes' : 'Phenotypes',
        '30_aggregate' : 'Summary statistics',
        '20_latest+demo' : 'Latest, demographics',
        '30_latest+demo+phenotypes' : 'Latest, demo., phenotypes',
        '20_latest+demo+aggregate' : 'Latest, demo., statistics',
        '60_all' : 'All'
    }
)

cv = cv.rename(
    columns = {
    'model' : 'Model',
    'dataset' : 'Feature Set',
    'roc_auc' : 'AUROC',
    'auprc' : 'AUPRC',
    'f1' : 'F1',
    'precision' : 'Precision',
    'recall' : 'Recall',
    }
)

Assess if phenotyping, summary statistics, and latest/demographic feature sets have statistically significant difference performance

x = cv[(cv['set'] == 'test')]

feature_sets = x['Feature Set'].unique()
metrics = ['AUROC', 'AUPRC', 'F1', 'Precision', 'Recall']

f_res = []
n_res = []
means = []

for metric in metrics:
    model_obs = []

    for i in range(len(feature_sets)):
        obs = x[(x['Feature Set'] == feature_sets[i])][metric].values.tolist()
        model_obs.append(obs)
        means.append([metric, feature_sets[i], x[(x['Feature Set'] == feature_sets[i])][metric].mean()])

    f, p = friedmanchisquare(*model_obs)
    f_res.append([metric, p])
    n_res.append([metric, pd.DataFrame(sp.posthoc_nemenyi_friedman(np.array(model_obs).T).values, columns=feature_sets, index=feature_sets)])

f_df = pd.DataFrame(f_res, columns=['Metric','p-value'])

means = pd.DataFrame(means, columns=['Metric', 'Feature Set', 'Mean'])

def format(entry):
    n = entry[1]
    mask = np.triu(np.ones(n.shape), k=1).astype(bool)
    n = n.where(mask)
    n = n.reset_index()
    n = n.rename(columns={'index': 'Feature Set 1'})
    n = n.melt(id_vars='Feature Set 1', var_name='Feature Set 2', value_name='p-value').dropna()
    n['Metric'] = entry[0]
    n = n.reset_index(drop=True)

    return n

format_n = []
for entry in n_res:
    format_n.append(format(entry))
n_df = pd.concat(format_n)

n_df = pd.merge(n_df, means, left_on=['Metric', 'Feature Set 1'], right_on=['Metric', 'Feature Set'])
n_df = pd.merge(n_df, means, left_on=['Metric', 'Feature Set 2'], right_on=['Metric', 'Feature Set'], suffixes=('_1', '_2'))
n_df = n_df.drop(columns=['Feature Set_1', 'Feature Set_2'])
n_df = n_df.rename(columns={'Mean_1': 'Mean of Feature Set 1', 'Mean_2': 'Mean of Feature Set 2'})

f_df[f_df['p-value'] < 0.05 / 5]

n_df[(n_df['p-value'] < 0.05)]

Compute if one model performed better than the others

x = cv[(cv['set'] == 'test')]

models = x['Model'].unique()
metrics = ['AUROC', 'AUPRC', 'F1', 'Precision', 'Recall']

f_res = []
n_res = []
means = []

for metric in metrics:
    model_obs = []

    for i in range(len(models)):
        obs = x[(x['Model'] == models[i])][metric].values.tolist()
        model_obs.append(obs)
        means.append([metric, models[i], x[(x['Model'] == models[i])][metric].mean()])


    f, p = friedmanchisquare(*model_obs)
    f_res.append([metric, p])
    n_res.append([metric, pd.DataFrame(sp.posthoc_nemenyi_friedman(np.array(model_obs).T).values, columns=models, index=models)])

f_df = pd.DataFrame(f_res, columns=['Metric','p-value'])
means = pd.DataFrame(means, columns=['Metric', 'Model', 'Mean'])


def format(entry):
    n = entry[1]
    mask = np.triu(np.ones(n.shape), k=1).astype(bool)
    n = n.where(mask)
    n = n.reset_index()
    n = n.rename(columns={'index': 'Model 1'})
    n = n.melt(id_vars='Model 1', var_name='Model 2', value_name='p-value').dropna()
    n['Metric'] = entry[0]
    n = n.reset_index(drop=True)

    return n

format_n = []
for entry in n_res:
    format_n.append(format(entry))
n_df = pd.concat(format_n)

n_df = pd.merge(n_df, means, left_on=['Metric', 'Model 1'], right_on=['Metric', 'Model'])
n_df = pd.merge(n_df, means, left_on=['Metric', 'Model 2'], right_on=['Metric', 'Model'], suffixes=('_1', '_2'))
n_df = n_df.drop(columns=['Model_1', 'Model_2'])
n_df = n_df.rename(columns={'Mean_1': 'Mean of Model 1', 'Mean_2': 'Mean of Model 2'})

f_df[f_df['p-value'] < 0.05 / 5]

n_df[(n_df['p-value'] < 0.05)]
